+++ src/hkimjp/datascript.clj
@@ -12,30 +12,32 @@
 
 (def storage nil)
 
-(def default-storage-url "jdbc:sqlite:resources/db.sqlite")
+(def default-storage-url "jdbc:sqlite:/tmp/db.sqlite")
 
 (defn- datasource
   [url]
+  (t/log! :info (str "datasource url: " url))
   (doto (org.sqlite.SQLiteDataSource.)
     (.setUrl url)))
 
 (defn- pooled-datasource
   [ds]
+  (t/log! :info "pooled-datasource")
   (storage-sql/pool ds {:max-conn 10 :max-idle-conn 4}))
 
 (time-literals.read-write/print-time-literals-clj!)
 
 (defn- sqlite-storage
-  [datasource]
-  (storage-sql/make datasource
+  [ds]
+  (t/log! :info "sqlite-stroage")
+  (storage-sql/make ds
                     {:dbtype     :sqlite
                      :freeze-str pr-str
                      :thaw-str   #(read-string {:readers rw/tags} %)}))
 
 (defn- make-storage [url]
-  (t/log! :info (str "make-storage " url))
-  (let [url (or url default-storage-url)
-        st (-> url
+  (t/log! :info (str "make-storage url: " url))
+  (let [st (-> url
                datasource
                pooled-datasource
                sqlite-storage)]
@@ -43,16 +45,19 @@
 
 (defn- create-conn
   ([schema]
-   (t/log! :info (str "create-conn " schema))
+   (t/log! :info (str "create-conn on-memory schema: " schema))
    (alter-var-root #'conn (constantly (d/create-conn schema))))
   ([schema storage]
-   (t/log! :info (str "create-conn " schema " with storage"))
+   (t/log! :info (str "create-conn with schema: "
+                      (if (nil? schema) "nil" schema)))
    (alter-var-root #'conn (constantly (d/create-conn schema storage)))))
 
 (defn- restore-conn [storage]
+  (t/log! :info (str "restore-conn"))
   (alter-var-root #'conn (constantly (d/restore-conn storage))))
 
 (defn- close-conn []
+  (t/log! :info "close-conn")
   (when (some? storage)
     (storage-sql/close storage)
     (alter-var-root #'storage (constantly nil)))
@@ -67,25 +72,47 @@
       (.exists (java.io.File. path)))
     (catch Exception _ false)))
 
+; (defn restore
+;   ([] (restore {:url default-storage-url}))
+;   ([{:keys [url]}]
+;    (t/log! :info (str "restore " url))
+;    (if (exist? url)
+;      (restore-conn (make-storage url))
+;      (throw (Exception. (str "does not exist " url))))))
+
 (defn restore
-  ([] (restore default-storage-url))
-  ([url]
-   (t/log! :info (str "restore " url))
-   (if (exist? url)
-     (restore-conn (make-storage url))
-     (throw (Exception. (str "does not exist " url))))))
+  ([] (restore {:url default-storage-url}))
+  ([{:keys [url] :as param}]
+   (t/log! :info (str "restore " param))
+   (if (nil? url)
+     (restore {:url param})
+     (if (exist? url)
+       (restore-conn (make-storage url))
+       (throw (Exception. (str "does not exist " url)))))))
 
 (defn start
   "If the :url argument is found and its value is nil, use the value of
    `default-storage-url` instead of nil.
    If you want an on-memory database, do not give the :url option.
    Use (restore) or (restore storage-url) when restoring."
-  ([] (create-conn nil nil))
+  ([]
+   (t/log! :info "on-memory datascript, no schema provided")
+   (create-conn nil nil))
   ([{:keys [schema url] :as params}]
-   (t/log! :info (str "start " params))
    (if (contains? params :url)
-     (create-conn schema {:storage (make-storage url)})
-     (create-conn schema nil))))
+     (let [url (or url default-storage-url)]
+       (t/log! :info (str "start with storage: " url))
+       (create-conn schema {:storage (make-storage url)}))
+     (do
+       (t/log! :info (str "start on-memory datascript schema: "
+                          (if (nil? schema) "nil" schema)))
+       (create-conn schema nil)))))
+
+(defn start-or-restore [{:keys [url] :as params}]
+  (t/log! :info (str "start-or-restore " params))
+  (if (exist? url)
+    (restore {:url url})
+    (start params)))
 
 (defn stop []
   (close-conn))
@@ -122,7 +149,7 @@
 ;;   `(d/q ~query @conn ~@inputs))
 
 (defn qq [query & inputs]
-  (t/log! :info (str "q " query))
+  (t/log! :debug (str "qq " query))
   (apply d/q query @conn inputs))
 
 (defn- supply-id [fact]
@@ -130,15 +157,20 @@
     fact
     (assoc fact :db/id -1)))
 
+(defn put! [fact]
+  (t/log! :debug (str "put! " fact))
+  (d/transact! conn [(supply-id fact)]))
+
 (defn puts! [facts]
-  (t/log! :info (str "puts " (abbrev facts)))
+  (t/log! :debug (str "puts! " (abbrev facts)))
   (d/transact! conn (map supply-id facts)))
 
 (defn pl
   ([eid] (pl ['*] eid))
   ([selector eid]
-   (t/log! :info (str "pull " selector " " eid))
+   (t/log! :debug (str "pull " selector " " eid))
    (d/pull @conn selector eid)))
 
 (defn et [id]
+  (t/log! :debug (str "et id " id))
   (d/entity @conn id))